requirements.txt
streamlit==1.31.0
openai==1.93.0
PyMuPDF==1.23.26
python-dotenv==1.0.1
tiktoken==0.5.2
numpy==1.24.3
pandas==2.0.3